rename to match rest of edge-endpoint
